+++ CHANGELOG.md
@@ -130,3 +130,4 @@ Please add a _short_ line describing the PR you make, if the PR implements a spe
 * Add `Access` column to show user if they have access or not ([#383](https://github.com/ScilifelabDataCentre/dds_cli/pull/383))
 * New `--mount-dir` option for `dds data put` where the `DataDelivery...` folders will be created if specified ([#393](https://github.com/ScilifelabDataCentre/dds_cli/pull/393))
 * New unit group command unit module ([#398](https://github.com/ScilifelabDataCentre/dds_cli/pull/398))
+* `--unit` option for Super Admins to list unit users ([#397](https://github.com/ScilifelabDataCentre/dds_cli/pull/397))